requirements.txt
langchain
langchain-community
langchain-openai

keybert
gtts

SentencePiece

sacremoses 
uvicorn
langserve

numpy
pandas

sentence_transformers

duckduckgo_search
bs4

transformers

fastapi
streamlit

python-dotenv